+++ app/(header-default)/dirigeants/_component/sections/summary.tsx
